yet another foolish mistake

+++ lobstero/cogs/editing.py
@@ -583,7 +583,7 @@ If you don't do any of that, Lobstero will search the previous few messages for 
 
         im = Image.open(result.data).convert("L")
         buffer = BytesIO()
-        to_write = numpy.array(im)
+        to_write = numpy.array(im, dtype=numpy.int8)
         wavfile.write(buffer, 10000, to_write)
         constructed_file = discord.File(fp=buffer, filename="audioimage.wav")
